rename Lb as InnerLb and move it to a seperate module file
