+++ libs/lb/lb-rs/src/lib.rs
@@ -23,44 +23,8 @@ pub mod io;
 pub mod model;
 pub mod service;
 pub mod subscribers;
-
-#[derive(Clone)]
-pub struct Lb {
-    pub config: Config,
-    pub keychain: Keychain,
-    pub db: LbDb,
-    pub docs: AsyncDocs,
-    pub search: SearchIndex,
-    pub client: Network,
-    pub events: EventSubs,
-    pub syncing: Arc<AtomicBool>,
-    pub status: StatusUpdater,
-}
-
-impl Lb {
-    #[instrument(level = "info", skip_all, err(Debug))]
-    pub async fn init(config: Config) -> LbResult<Self> {
-        logging::init(&config)?;
-
-        let db = CoreDb::init(db_rs::Config::in_folder(&config.writeable_path))
-            .map_err(|err| LbErrKind::Unexpected(format!("{:#?}", err)))?;
-        let keychain = Keychain::from(db.account.get());
-        let db = Arc::new(RwLock::new(db));
-        let docs = AsyncDocs::from(&config);
-        let client = Network::default();
-        let search = SearchIndex::default();
-        let status = StatusUpdater::default();
-        let syncing = Arc::default();
-        let events = EventSubs::default();
-
-        let result = Self { config, keychain, db, docs, client, search, syncing, events, status };
-
-        result.setup_search();
-        result.setup_status().await?;
-
-        Ok(result)
-    }
-}
+pub mod inner_lb;
+pub use inner_lb::InnerLb;
 
 pub fn get_code_version() -> &'static str {
     env!("CARGO_PKG_VERSION")
@@ -69,19 +33,6 @@ pub fn get_code_version() -> &'static str {
 pub static DEFAULT_API_LOCATION: &str = "https://api.prod.lockbook.net";
 pub static CORE_CODE_VERSION: &str = env!("CARGO_PKG_VERSION");
 
-use crate::io::CoreDb;
-use crate::service::logging;
-use db_rs::Db;
-use io::docs::AsyncDocs;
-use io::network::Network;
-use io::LbDb;
-use model::core_config::Config;
-pub use model::errors::{LbErrKind, LbResult};
-use service::events::EventSubs;
-use service::keychain::Keychain;
-use std::sync::atomic::AtomicBool;
-use std::sync::Arc;
-use subscribers::search::SearchIndex;
-use subscribers::status::StatusUpdater;
-use tokio::sync::RwLock;
 pub use uuid::Uuid;
+pub use model::errors::{LbErrKind, LbResult};
+pub use model::core_config::Config;
