+++ libs/lb/lb-rs/src/service/activity.rs
@@ -1,6 +1,6 @@
 use crate::model::errors::LbResult;
 use crate::model::tree_like::TreeLike;
-use crate::Lb;
+use crate::InnerLb as Lb;
 use serde::Deserialize;
 use serde::Serialize;
 use std::cmp;
