+++ libs/lb/lb-rs/src/service/file.rs
@@ -5,7 +5,7 @@ use crate::model::file_metadata::{FileType, Owner};
 use crate::model::filename::MAX_FILENAME_LENGTH;
 use crate::model::symkey;
 use crate::model::tree_like::TreeLike;
-use crate::Lb;
+use crate::InnerLb as Lb;
 use std::iter;
 use uuid::Uuid;
