+++ libs/lb/lb-rs/src/service/keychain.rs
@@ -9,7 +9,7 @@ use crate::{
         crypto::AESKey,
         errors::{LbErrKind, LbResult},
     },
-    Lb,
+    InnerLb as Lb,
 };
 use libsecp256k1::PublicKey;
 use tokio::sync::OnceCell;
