+++ libs/lb/lb-rs/src/service/path.rs
@@ -2,7 +2,7 @@ use crate::model::errors::{LbErrKind, LbResult};
 use crate::model::file::File;
 use crate::model::path_ops::Filter;
 use crate::model::tree_like::TreeLike;
-use crate::Lb;
+use crate::InnerLb as Lb;
 use uuid::Uuid;
 
 impl Lb {
