+++ libs/lb/lb-rs/src/service/sync.rs
@@ -18,7 +18,7 @@ use crate::model::tree_like::TreeLike;
 use crate::model::work_unit::WorkUnit;
 use crate::model::{clock, svg};
 use crate::model::{symkey, ValidationFailure};
-use crate::Lb;
+use crate::InnerLb as Lb;
 pub use basic_human_duration::ChronoHumanDuration;
 use futures::stream;
 use futures::StreamExt;
