+++ libs/lb/lb-rs/src/service/usage.rs
@@ -3,7 +3,7 @@ use crate::model::errors::LbResult;
 use crate::model::file_like::FileLike;
 use crate::model::tree_like::TreeLike;
 use crate::model::usage::{bytes_to_human, get_usage};
-use crate::Lb;
+use crate::InnerLb as Lb;
 use serde::Serialize;
 use std::collections::HashMap;
 use uuid::Uuid;
